+++ server/src/main/scala/org/http4s/server/middleware/HSTS.scala
@@ -2,9 +2,10 @@ package org.http4s
 package server
 package middleware
 
-import cats._
+import cats.Functor
 import cats.data.Kleisli
 import org.http4s.headers.`Strict-Transport-Security`
+
 import scala.concurrent.duration._
 
 /** [[Middleware]] to add HTTP Strict Transport Security (HSTS) support adding
@@ -17,23 +18,23 @@ object HSTS {
     includeSubDomains = true,
     preload = false)
 
-  def apply[F[_]: Functor](@deprecatedName('service) routes: HttpRoutes[F]): HttpRoutes[F] =
+  def apply[F[_]: Functor, A, G[_]: Functor](
+      @deprecatedName('service) routes: Kleisli[F, A, Response[G]]): Kleisli[F, A, Response[G]] =
     apply(routes, defaultHSTSPolicy)
 
-  def apply[F[_]: Functor](
-      @deprecatedName('service) routes: HttpRoutes[F],
-      header: `Strict-Transport-Security`): HttpRoutes[F] = Kleisli { req =>
-    routes.map(_.putHeaders(header)).apply(req)
+  def apply[F[_]: Functor, A, G[_]: Functor](
+      @deprecatedName('service) http: Kleisli[F, A, Response[G]],
+      header: `Strict-Transport-Security`): Kleisli[F, A, Response[G]] = Kleisli { req =>
+    http.map(_.putHeaders(header)).apply(req)
   }
 
-  def unsafeFromDuration[F[_]: Functor](
-      @deprecatedName('service) routes: HttpRoutes[F],
+  def unsafeFromDuration[F[_]: Functor, A, G[_]: Functor](
+      @deprecatedName('service) http: Kleisli[F, A, Response[G]],
       maxAge: FiniteDuration = 365.days,
       includeSubDomains: Boolean = true,
-      preload: Boolean = false): HttpRoutes[F] = {
+      preload: Boolean = false): Kleisli[F, A, Response[G]] = {
     val header = `Strict-Transport-Security`.unsafeFromDuration(maxAge, includeSubDomains, preload)
-
-    apply(routes, header)
+    apply(http, header)
   }
 
 }